requirements.txt
Django==1.4.1
South==0.7.4
Pillow==1.7.7
django-tastypie==0.9.11
django_compressor==1.2
cssmin==0.1.4
jsmin==2.0.2
psycopg2==2.4.5
django-extensions==0.9
elementtree==1.2.7-20070827-preview
boto==2.3.0
django-storages==1.1.4
gunicorn==0.14.2
raven==1.8.4
django-celery==3.0.6
celery==3.0.8
